Separated out the new AnswerTest class from the old code by putting the new code into a module called answer_tests.py.
